requirements.txt
streamlit>=1.29.0
snowflake-connector-python>=3.6.0
openai>=1.12.0
pandas>=2.2.0
python-dotenv>=1.0.0
streamlit-ace>=0.1.1
openpyxl>=3.1.2
pyarrow>=10.0.0,<19.0.0 